+++ EMSTwigListBundle/Resources/config/config.yml
@@ -1,0 +1,1 @@
+ems_twig_list :    templates:        - {resource: 'EMSTwigListBundle', path: 'Resources/views/Pages'}    app_enabled: false    app_base_path : #['Resources/views/default']
